# sectorstorage: Make trySched less very slow