This doesn't make it any less O(n^2), just runs the slow part in parallel, so it's acceptably fast at more scales

Before:
```
BenchmarkTrySched/200w-400q-32         	       1	21105421605 ns/op (21s)
```

After:
```
BenchmarkTrySched/200w-400q-32         	       9	 130080438 ns/op (130ms)
```